flask_api/requirements.txt
Flask
firebase-admin
Flask-CORS
python-dotenv